implemented: order related stuff and scheduler for cancel expired orders

+++ src/main/resources/schema.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS users (
 CREATE TABLE IF NOT EXISTS products (
     id BIGSERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
-    price NUMERIC(10, 2) NOT NULL CHECK (price >= 0),
+    price DOUBLE PRECISION NOT NULL CHECK (price >= 0),
     stock_quantity INT NOT NULL CHECK (stock_quantity >= 0),
     created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP
 );
@@ -28,7 +28,7 @@ CREATE TABLE IF NOT EXISTS order_item (
     order_id BIGINT NOT NULL REFERENCES orders(id) ON DELETE CASCADE,
     product_id BIGINT NOT NULL REFERENCES products(id) ON DELETE RESTRICT,
     quantity INT NOT NULL CHECK (quantity > 0),
-    price NUMERIC(10, 2) NOT NULL CHECK (price >= 0)
+    price DOUBLE PRECISION NOT NULL CHECK (price >= 0)
 );
 
 CREATE TABLE IF NOT EXISTS refresh_token(
